+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.0.2 (SNAPSHOT)
-* Add mock http request and response.
+* Add mock http request and response
+* Add mock process platform
+* Add Windows support for cli readConfigFileSync, should identify home directory properly now 
 
 ### 0.0.1
 * Initial version